chore: move title to legend
