+++ __tests__/integration/components/index.ts
@@ -1,7 +1,6 @@
 export * from './axis';
 export * from './legend';
 export * from './navigator';
-export * from './title';
 export * from './indicator';
 export * from './crosshair';
 export * from './bugs';
