+++ __tests__/integration/components/legend/index.ts
@@ -75,3 +75,4 @@ export { CategoryItemsPoptip2 } from './category-items-poptip-2';
 export { CategoryItemsPoptip3 } from './category-items-poptip-3';
 export { CategoryItemsPoptip4 } from './category-items-poptip-4';
 export { CategoryItemsPoptip5 } from './category-items-poptip-5';
+export { TitleDemo } from './title';
