--- __tests__/integration/components/title/index.ts
@@ -1,1 +1,0 @@
-export * from './title';
