+++ src/ui/legend/category.ts
@@ -1,8 +1,8 @@
 import { Component } from '../../core';
 import type { Group } from '../../shapes';
 import { BBox, select, Selection, splitStyle, subStyleProps } from '../../util';
-import type { TitleStyleProps } from '../title';
-import { getBBox, Title } from '../title';
+import type { TitleStyleProps } from './title';
+import { getBBox, Title } from './title';
 import { CategoryItems } from './category/items';
 import { CATEGORY_DEFAULT_OPTIONS, CLASS_NAMES } from './constant';
 import type { CategoryOptions, CategoryStyleProps } from './types';
