+++ src/util/index.ts
@@ -39,3 +39,4 @@ export * from './transpose';
 export * from './traverse';
 export * from './visibility';
 export * from './replace-children';
+export * from './pase-position';
